+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.2
+
+- Added `resolveBoth` for `FutureOr` and `Future`.
+- Add `Future` and `FutureOr` arithmetic operators.
+- Added Benchmarks:
+  - `async_extension_benchmark.dart`
+  - `async_extension_benchmark2.dart`
+
 ## 1.0.1
 
 - Adjusted `pubspec.yaml` description.